+++ pom.xml
@@ -300,8 +300,9 @@
       <version>${jetty.version}</version>
     </dependency>
 
+
     <dependency>
-      <groupId>org.eclipse.jetty.alpn</groupId>
+      <groupId>org.eclipse.jetty</groupId>
       <artifactId>jetty-alpn-server</artifactId>
       <version>${jetty.version}</version>
     </dependency>
