+++ src/java/winstone/LocalStrings.properties
@@ -150,3 +150,5 @@ HttpsConnectorFactory.LoadPrivateKey=\
   Using the --httpsPrivateKey/--httpsCertificate options currently relies on unsupported APIs in the Oracle JRE.\n\
   Please use --httpsKeyStore and related options instead.
 HttpsConnectorFactory.LoadPrivateKeyError=Cannot load private key; try using a Java keystore instead.
+
+Http2ConnectorFactory.FailStart.ALPN=Fail to start ALPN
